Coding - Cast RWGltf_GltfRootElement to int for array size/subscript
macOS Clang reports value-not-implicitly-convertible-to-unsigned-long
for the static THE_ROOT_NAMES array size and the return subscript;
wrap the enum class value with static_cast<int>() at three sites
(RWGltf_GltfRootElement.hxx array declaration and subscript,
RWGltf_GltfJsonParser.hxx member array size).

The secondary ParseStream errors in RWGltf_CafReader are a downstream
instantiation failure that resolves once the parser header compiles.

+++ src/DataExchange/TKDEGLTF/RWGltf/RWGltf_GltfJsonParser.hxx
@@ -522,8 +522,8 @@ protected:
   // clang-format on
 
 #ifdef HAVE_RAPIDJSON
-  GltfElementMap
-    myGltfRoots[RWGltf_GltfRootElement::RWGltf_GltfRootElement_NB]; //!< glTF format root elements
+  GltfElementMap myGltfRoots[static_cast<int>(
+    RWGltf_GltfRootElement::RWGltf_GltfRootElement_NB)]; //!< glTF format root elements
 #endif
 };
